refactor(ui): split collection registry into focused modules
